Extract Box interface, implement FileBox
