+++ src/test/java/pokemanager/ExitCommandTest.java
@@ -9,7 +9,7 @@ import static org.junit.Assert.*;
 public class ExitCommandTest {
     private final ByteArrayOutputStream out = new ByteArrayOutputStream();
     private final PrintStream pw = new PrintStream(out);
-    private final BoxSpy box = new BoxSpy(new ByteArrayInputStream("".getBytes()));
+    private final BoxSpy box = new BoxSpy();
     private final ExitCommand ec = new ExitCommand();
     private final InputStream in = new ByteArrayInputStream("".getBytes());
     private final AppSpy app = new AppSpy(in, pw, box, "", null);
