+++ src/test/java/pokemanager/RetrieveCommandTest.java
@@ -11,7 +11,7 @@ public class RetrieveCommandTest {
 
     private final ByteArrayOutputStream out = new ByteArrayOutputStream();
     private final PrintStream pw = new PrintStream(out);
-    private final BoxSpy box = new BoxSpy(new ByteArrayInputStream("".getBytes()));
+    private final BoxSpy box = new BoxSpy();
     private final RetrieveCommand rc = new RetrieveCommand();
     private final InputStream in = new ByteArrayInputStream("".getBytes());
     private final App app = new App(in, pw, box, "", null);
