+++ src/test/java/pokemanager/SpeciesCommandTest.java
@@ -11,7 +11,7 @@ public class SpeciesCommandTest {
 
     private final ByteArrayOutputStream out = new ByteArrayOutputStream();
     private final PrintStream pw = new PrintStream(out);
-	private final BoxSpy box = new BoxSpy(new ByteArrayInputStream("".getBytes()));
+	private final BoxSpy box = new BoxSpy();
     private final InputStream in = new ByteArrayInputStream("".getBytes());
     private final File tempFile = File.createTempFile("temp-", "-testfile");
     private final App app = new App(in, pw, box, tempFile.toString(), new HttpGetRequesterSpy());
